bookmarking/requirements.txt
beautifulsoup4==4.12.2
Django==4.2.4
djangorestframework==3.14.0
djoser==2.2.0
drf-extra-fields==3.7.0
drf-yasg==1.21.7
psycopg2-binary
